refac: transforming selectAll def to a file .py
